project/requirements.txt
asyncpg==0.29.0
black==24.3.0
email-validator==2.0.0
fastapi==0.110.0
fastapi-pagination==0.12.24
firebase-admin==6.5.0
flake8==7.0.0
gunicorn==21.2.0
httpx==0.27.0
isort==5.13.2
jinja2
passlib==1.7.4
pydantic-settings==2.2.1
pyjwt==2.7.0
pytest==8.1.1
python-multipart==0.0.9
requests==2.31.0
tortoise-orm==0.20.0
uvicorn==0.29.0